requirements.txt
flask
requests
gunicorn
python-dotenv
flask-jwt-extended
flasgger
